# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Микросервис для работы с балансом пользователей",
        "title": "Balance Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@balance-service.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/balance": {
            "patch": {
                "description": "Зачисление средств на баланс пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Пополнение баланса",
                "parameters": [
                    {
                        "description": "Данные для пополнения баланса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_user.AddBalanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные пользователя с обновленным балансом",
                        "schema": {
                            "$ref": "#/definitions/internal_user.User"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/internal_user.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_user.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/balance/{user_id}": {
            "get": {
                "description": "Возвращает информацию о балансе пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Получение баланса",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные пользователя",
                        "schema": {
                            "$ref": "#/definitions/internal_user.User"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID пользователя",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/history": {
            "get": {
                "description": "Возвращает историю транзакций пользователя с пагинацией и сортировкой",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "История транзакций",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Номер страницы (по умолчанию 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страницы (по умолчанию 20)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поле для сортировки (amount/date)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Порядок сортировки (asc/desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "История транзакций",
                        "schema": {
                            "$ref": "#/definitions/internal_history.TransactionHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report": {
            "get": {
                "description": "Генерирует CSV отчет по выручке за указанный месяц",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Генерация отчета",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Год (например: 2024)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Месяц (1-12)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ссылка на отчет",
                        "schema": {
                            "$ref": "#/definitions/internal_stat.ReportResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reserve": {
            "post": {
                "description": "Резервирует средства на балансе пользователя для указанной услуги",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Резервирование средств",
                "parameters": [
                    {
                        "description": "Данные для резервирования",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_reservations.ReserveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат резервирования",
                        "schema": {
                            "$ref": "#/definitions/internal_reservations.ReserveResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные для резервирования",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reserve/cancel/{reservation_id}": {
            "patch": {
                "description": "Отменяет ранее созданное резервирование средств",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Отмена резервирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID резервирования",
                        "name": "reservation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная отмена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка отмены",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reserve/confirm/{reservation_id}": {
            "patch": {
                "description": "Подтверждает ранее созданное резервирование средств",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Подтверждение резервирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID резервирования",
                        "name": "reservation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное подтверждение",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка подтверждения",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transfer": {
            "post": {
                "description": "Переводит указанную сумму от одного пользователя другому",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfer"
                ],
                "summary": "Перевод средств",
                "parameters": [
                    {
                        "description": "Данные для перевода",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_user.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат перевода",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректные данные для перевода",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "Создает нового пользователя с начальным балансом 0",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создание пользователя",
                "parameters": [
                    {
                        "description": "Данные для создания пользователя",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_user.UserCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь создан",
                        "schema": {
                            "$ref": "#/definitions/internal_user.User"
                        }
                    },
                    "400": {
                        "description": "Ошибка создания пользователя",
                        "schema": {
                            "$ref": "#/definitions/internal_user.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_history.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "from_user_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "related_id": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "to_user_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_history.TransactionHistoryResponse": {
            "type": "object",
            "properties": {
                "current_page": {
                    "description": "Текущая страница",
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "description": "Общее количесво",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "Общее количество страниц",
                    "type": "integer"
                },
                "transactions": {
                    "description": "Слайс представляющий информацию",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_history.Transaction"
                    }
                }
            }
        },
        "internal_reservations.ReserveRequest": {
            "description": "Запрос для резервирования средств перед оплатой",
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Сумма которая резервируется",
                    "type": "number"
                },
                "order_id": {
                    "description": "ID заказа",
                    "type": "integer"
                },
                "service_id": {
                    "description": "ID услуги",
                    "type": "integer"
                },
                "user_id": {
                    "description": "ID пользователя",
                    "type": "integer"
                }
            }
        },
        "internal_reservations.ReserveResponse": {
            "description": "Ответ с данными о резервировании средств",
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Сумма которая резервируется",
                    "type": "number"
                },
                "id": {
                    "description": "ID резервации",
                    "type": "integer"
                },
                "new_balance": {
                    "description": "Обновленый баланс пользователя после списания средств",
                    "type": "number"
                },
                "order_id": {
                    "description": "ID заказа",
                    "type": "integer"
                },
                "service_id": {
                    "description": "ID услуги",
                    "type": "integer"
                },
                "status": {
                    "description": "Статус",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID пользователя",
                    "type": "integer"
                }
            }
        },
        "internal_stat.ReportResponse": {
            "type": "object",
            "properties": {
                "file_name": {
                    "description": "Название файла",
                    "type": "string"
                },
                "message": {
                    "description": "Сообщение",
                    "type": "string"
                },
                "report_url": {
                    "description": "Путь",
                    "type": "string"
                }
            }
        },
        "internal_user.AddBalanceRequest": {
            "description": "Запрос для пополнения баланса пользователя",
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Сумма пополнения",
                    "type": "number"
                },
                "user_id": {
                    "description": "ID пользователя",
                    "type": "integer"
                }
            }
        },
        "internal_user.ErrorResponse": {
            "description": "Стандартный ответ при возникновении ошибки",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Описание ошибки"
                }
            }
        },
        "internal_user.TransferRequest": {
            "description": "Запрос для перевода средств между пользователями",
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Сумма пополнения",
                    "type": "number"
                },
                "from_user_id": {
                    "description": "Имя пользователя отправителя",
                    "type": "integer"
                },
                "to_user_id": {
                    "description": "Имя пользователя получателя",
                    "type": "integer"
                }
            }
        },
        "internal_user.User": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_user.UserCreateRequest": {
            "description": "Запрос для создания нового пользователя",
            "type": "object",
            "properties": {
                "name": {
                    "description": "Имя пользователя",
                    "type": "string"
                }
            }
        }
    }
}